requirements.txt
annotated-types==0.7.0
Authlib==1.3.2
bandit==1.7.9
boto3==1.35.19
botocore==1.35.19
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
coverage==7.6.1
cryptography==43.0.1
dparse==0.6.4b0
filelock==3.12.4
flake8==7.1.1
idna==3.10
iniconfig==2.0.0
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mccabe==0.7.0
mdurl==0.1.2
packaging==24.1
pbr==6.1.0
pluggy==1.5.0
psutil==6.0.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
pyflakes==3.2.0
Pygments==2.18.0
pytest==8.3.3
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rich==13.8.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.2
safety==3.2.7
safety-schemas==0.0.5
shellingham==1.5.4
six==1.16.0
stevedore==5.3.0
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
